Check for package replacements in docker install script

+++ build_files/install-docker.sh
@@ -22,6 +22,16 @@ dnf5 config-manager addrepo --from-repofile="https://download.docker.com/linux/f
 # Disable repo
 dnf5 config-manager setopt docker-ce-stable.enabled=0
 
+### Check if base image packages are being replaced
+# Dry run
+dnf5 -y install --enable-repo="docker-ce-stable" --setopt=tsflags=test "${docker_pkgs[@]}" 2>&1 | tee /tmp/dryrun.log
+
+# Check log for upgrading and downgrading
+if grep -qE '^(Upgrading|Downgrading):' /tmp/dryrun.log; then
+	echo "::notice::Detected package replacements. Aborting build."
+	exit 1
+fi
+
 # Temporarily enable repo and install Docker packages
 dnf5 install -y --enable-repo="docker-ce-stable" "${docker_pkgs[@]}" || {
     # Use test packages if docker pkgs is not available for f42
